requirements.txt
Django>=3.1.0,<3.2.0
djangorestframework>=3.11.1,<3.12.0
psycopg2>=2.8.5,<2.9.0
pylint-django>=2.3.0,<2.4.0

flake8>=3.8.3,<3.9.0
